% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_sim-data.R
\docType{data}
\name{dat_sim}
\alias{dat_sim}
\title{Simluated data to be used with \code{glmmr}}
\format{
The data object \code{dat_sim} is a \code{data.frame} containing 250 binary outcomes, a group variable (id) and a covariate (time).
}
\usage{
data(dat_sim)
}
\description{
The data object \code{dat_sim} contains 250 simulated datasets for demonstrating a mixed logistic regression model.
}
\keyword{datasets}
