% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Utilities.R
\name{GetL}
\alias{GetL}
\title{GetL}
\usage{
GetL(Z, Q)
}
\arguments{
\item{Z}{Z matrix}

\item{Q}{dimension of the correlation matrix}
}
\value{
\code{GetL} returns a matrix of size Q
}
\description{
\code{GetL} is a function to computes the Cholesky factor for a correlation matrix from the intermediate matrix Z.
}
\examples{
###Test function
l <- -0.34
Z <- GetZ(l, 2)
L <- GetL(Z, 2)

}
