% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Utilities.R
\name{GetZ}
\alias{GetZ}
\title{GetZ}
\usage{
GetZ(l, Q)
}
\arguments{
\item{l}{l vector of unconstrained parameters}

\item{Q}{dimension of the correlation matrix}
}
\value{
\code{GetZ} returns a matrix of size Q
}
\description{
\code{GetZ} is a function to compute the intermediate step for the Cholesky factor of a correlation matrix.
}
\examples{
###Test function
l <- -0.34
Z <- GetZ(l, 2)
L <- GetL(Z, 2)

}
