\name{glmmr-package}
\alias{glmmr-package}
\alias{glmmr}
\docType{package}
\title{
\packageTitle{glmmr}
}
\description{
\packageDescription{glmmr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{glmmr}
\packageIndices{glmmr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{glmmr}

Maintainer: \packageMaintainer{glmmr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
