% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmr-package.R
\docType{package}
\name{glmmr}
\alias{glmmr}
\title{glmmr}
\description{
\code{glmmr} is a package for Bayesian generalized linear mixed models using stochastic gradient Markov chain Monte Carlo.
}
\author{
Samuel I. Berchuck \email{sib2@duke.edu}
}
