% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Utilities.R
\name{is.glmmr}
\alias{is.glmmr}
\title{is.glmmr}
\usage{
is.glmmr(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
\code{is.glmmr} returns a logical, depending on whether the object is of class \code{\link{glmmr}}.
}
\description{
\code{#' is.glmmr} is a general test of an object being interpretable as a
\code{\link{glmmr}} object.
}
\details{
The \code{\link{glmmr}} class is defined as the regression object that
 results from the \code{\link{glmmr}} regression function.
}
\examples{
###Load pre-computed results
data(reg_glmmr)

###Test function
is.glmmr(reg_glmmr)

}
